more unit cleanup

+++ src/filter.rs
@@ -22,11 +22,11 @@ impl<S> LowPass<S> {
         frequency: u64,
         decimate: u64,
         original_sample_rate: u64,
-        band: f32,
+        band: f64,
     ) -> Self {
         let cutoff = cutoff_from_frequency(frequency as f64, original_sample_rate);
 
-        let filter = lowpass_filter(cutoff as f32, band);
+        let filter = lowpass_filter(cutoff as f32, band as f32);
         LowPass {
             inner,
             filter,
